Refactor: Rename generateGraphs.py to generate_graphs.py and update imports
